CLDN-1118: Grabbing access token and login server name to connect to acr and docker registriesto get the trino image

+++ run-trino-env.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 trinoCloneDir=${1:-~/repositories/trino}
 
@@ -20,12 +20,22 @@ else
 fi
 
 echo "Logging in to Azure Container Registry"
-az acr login --name uchimera >/dev/null
+#read -r LOGIN_SERVER ACCESS_TOKEN <<< $(az acr login -n uchimera --expose-token --query '[loginServer, accessToken]' -otsv 2>/dev/null)
+read -r ACCESS_TOKEN LOGIN_SERVER <<< $(az acr login -n uchimera --expose-token  -otsv 2>/dev/null)
 ACR_LOGIN_STATUS=$?
 if [ ${ACR_LOGIN_STATUS} -eq 0 ]; then
-  mkdir -p ~/workspaces/trino
-  echo "Starting IDE"
-  docker run --rm --pull always -p 8887:8887 --mount type=bind,source="${HOME}/workspaces/trino",target=/home/projector-user --mount type=bind,source="${trinoCloneDir}",target=/workspace/trino -it uchimera.azurecr.io/cccs/dev/projector-intellij-ce:trino
+  echo "Logging in to Docker Registry"
+#  docker login ${LOGIN_SERVER} -u 00000000-0000-0000-0000-000000000000 -p ${ACCESS_TOKEN}
+  docker login $(echo ${LOGIN_SERVER} | tr -d '\r') -u 00000000-0000-0000-0000-000000000000 -p ${ACCESS_TOKEN}
+  DOCKER_LOGIN_STATUS=$?
+  if [ ${DOCKER_LOGIN_STATUS} -eq 0 ]; then
+    mkdir -p ~/workspaces/trino
+    echo "Starting IDE"
+    docker run --rm --pull always -p 8887:8887 --mount type=bind,source="${HOME}/workspaces/trino",target=/home/projector-user --mount type=bind,source="${trinoCloneDir}",target=/workspace/trino -it uchimera.azurecr.io/cccs/dev/projector-intellij-ce:trino
+  else
+  echo "Docker login failed"
+  exit 1
+  fi
 else
   echo "Azure Container Registry login failed"
   exit 1
